archlinux: Missing packages for using archlinux as a NetVM

+++ scripts_archlinux/09_cleanup.sh
@@ -8,7 +8,15 @@ sudo mount $CACHEDIR/root-image.fs mnt_archlinux_dvd
 echo "--> Starting cleanup actions"
 # Remove unused packages and their dependencies (make dependencies)
 cleanuppkgs=`sudo ./mnt_archlinux_dvd/usr/bin/arch-chroot $INSTALLDIR pacman -Qdt | cut -d " " -f 1`
-sudo ./mnt_archlinux_dvd/usr/bin/arch-chroot $INSTALLDIR pacman --noconfirm -Rsc $cleanuppkgs
+echo "--> Packages that can be cleaned up: $cleanuppkgs"
+if [ -n "$cleanuppkgs" ] ; then
+	sudo ./mnt_archlinux_dvd/usr/bin/arch-chroot $INSTALLDIR pacman --noconfirm -Rsc $cleanuppkgs
+fi
+
+# Remove non required linux kernel
+echo "--> Cleaning up linux kernel"
+sudo ./mnt_archlinux_dvd/usr/bin/arch-chroot $INSTALLDIR pacman --noconfirm -Rsc linux
+sudo ./mnt_archlinux_dvd/usr/bin/arch-chroot $INSTALLDIR pacman --noconfirm -S linux-firmware
 
 # Clean pacman cache
 sudo ./mnt_archlinux_dvd/usr/bin/arch-chroot $INSTALLDIR pacman --noconfirm -Scc
